backup_complex_version/requirements_complex.txt
# Core ML and Vision
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0

# Visualization and Reporting
matplotlib>=3.7.0

# Progress tracking
tqdm>=4.65.0

# For Excel export
openpyxl>=3.1.0